requirements.txt
apipkg==1.5
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
clize==4.0.3
cryptography==2.3.1
decorator==4.3.0
docutils==0.14
execnet==1.5.0
idna==2.7
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
more-itertools==6.0.0
nexmo==2.2.0
numpy==1.15.1
od==1.0
pandas==0.23.4
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
pkg-resources==0.0.0
pluggy==0.9.0
prompt-toolkit==2.0.7
ptyprocess==0.6.0
py==1.8.0
py-dotenv==0.1
pycparser==2.19
pydotenv==0.0.7
Pygments==2.3.1
PyJWT==1.6.4
pytest==4.3.0
pytest-forked==1.0.2
pytest-xdist==1.26.1
python-dateutil==2.7.3
python-dotenv==0.9.1
pytz==2018.5
requests==2.19.1
sigtools==2.0.2
six==1.11.0
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
xlrd==1.1.0
